+++ lib/semantic_release/configuration.rb
@@ -2,10 +2,11 @@
 
 module SemanticRelease
   class Configuration
-    attr_accessor :semver_file
+    attr_accessor :semver_file, :disable_rubygems_message
 
     def initialize
       @semver_file = ".semver"
+      @disable_rubygems_message = false
     end
   end
 end